requirements.txt
asgiref==3.8.1
Django==5.1.1
django-crispy-forms==2.3
django-extensions==4.1
django-widget-tweaks==1.5.0
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
